Initial commit of the HarkerRoboLib! (see desc.)
So this is my idea for a library that can be used from year to year
saving rewrite (and/or copy paste) time. Currently it only contains
what we’ve written so far, but it would be nice for us to keep adding
to it. As WPILibJ uses a custom build file, I’m not entirely sure how
to add it as an actual library, and instead have copied it over as a
separate package into RobotCode2015; however, if anyone could think of
a way to do so, that would be awesome. That’s all for now, folks.